External event API for send/recv implemented.
This was the main blocker, I think, for the nanomsg legacy compat
shim.  Now that we have this, it should be relatively straight-forward
to implement the legacy nanomsg API, including the SENDFD, RECVFD thing.

+++ tests/convey.c
@@ -705,7 +705,7 @@ convey_vlogf(struct convey_log *log, const char *fmt, va_list va, int addnl)
 {
 	/* Grow the log buffer if we need to */
 	while ((log->log_size - log->log_length) < 256) {
-		int newsz = log->log_size + 2000;
+		size_t newsz = log->log_size + 2000;
 		char *ptr = malloc(newsz);
 		if (ptr == NULL) {
 			return;
